+++ frontend/app/forms/MemberForm.scala
@@ -108,7 +108,6 @@ object MemberForm {
     lazy val zuoraAccountAddress = billingAddress.getOrElse(deliveryAddress)
   }
 
-  case class FeedbackForm(category: FeedbackType, page: String, feedback: String, name: String, email: String)
 
   val abTestFormatter: Formatter[JsValue] = new Formatter[JsValue] {
     override def bind(key: String, data: Map[String, String]): Either[Seq[FormError],JsValue] = {
@@ -143,21 +142,10 @@ object MemberForm {
       Map(key -> value.alpha2)
   }
 
-  implicit val feedbackTypeFormatter: Formatter[FeedbackType] = new Formatter[FeedbackType] {override def unbind(key: String, value: FeedbackType) = Map(key -> value.slug)
-
-    override def bind(key: String, data: Map[String, String]):Either[Seq[FormError],FeedbackType] = {
-      val feedbackType = data.get(key)
-      lazy val formError = FormError(key, s"Cannot find a feedback type of ${feedbackType.getOrElse("")}")
-      feedbackType
-        .flatMap(FeedbackType.fromSlug)
-        .toRight[Seq[FormError]](Seq(formError))
-    }
-  }
 
 
   private val productFeature = of[Set[FeatureChoice]] as productFeaturesFormatter
 
-  private val feedbackType = of[FeedbackType] as feedbackTypeFormatter
 
   private val country: Mapping[String] =
     text.verifying { code => CountryGroup.countryByCode(code).isDefined }
@@ -218,13 +206,7 @@ object MemberForm {
     "payPalBaid" -> optional(text)
   )(PaymentForm.apply)(PaymentForm.unapply)
 
-  val feedbackMapping: Mapping[FeedbackForm] =   mapping(
-    "category" -> feedbackType,
-    "page" -> text,
-    "feedback" -> nonEmptyText,
-    "name" -> nonEmptyText,
-    "email" -> email
-  )(FeedbackForm.apply)(FeedbackForm.unapply)
+
 
   val friendJoinForm: Form[FriendJoinForm] = Form(
     mapping(
@@ -282,15 +264,7 @@ object MemberForm {
     )(PaidMemberChangeForm.apply)(PaidMemberChangeForm.unapply)
   )
 
-  val feedbackForm: Form[FeedbackForm] = Form(
-    mapping(
-      "category" -> feedbackType,
-      "page" -> text,
-      "feedback" -> nonEmptyText,
-      "name" -> nonEmptyText,
-      "email" -> email
-    )(FeedbackForm.apply)(FeedbackForm.unapply)
-  )
+
 
   val supportForm: Form[SupportForm] = Form(
     mapping(
